+++ NEWS.md
@@ -1,6 +1,9 @@
-# TreeTools 1.7.3.9000 (development) #
+# TreeTools 1.7.3.9001 (development) #
 
 - Improve `DropTip()` speed when branch lengths are present.
+
+- Support > 32767 trees in `Consensus()` 
+  ([#127](https://github.com/ms609/TreeTools/issues/127))
 
 
 # TreeTools 1.7.3 #